% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerLRT.R
\name{powerLRT}
\alias{powerLRT}
\title{Estimate the Power for a Likelihood Ratio Test}
\usage{
powerLRT(diffLL, diffDF, alpha = 0.05)
}
\arguments{
\item{diffLL}{The NCP in a given LRT}

\item{diffDF}{The difference in the degrees of freedom}

\item{alpha}{The desired alpha. Default = 0.05}
}
\value{
The power to reject the Null Hypothesis.
}
\description{
This function estimates the power to reject the Null Hypothesis in a Likelihood Ratio Test
for two nested models fit using Maximum Likelihood Estimation.
The function requires as input the desired alpha (\code{alpha}; default = 0.05),
the non-centrality parameter (\code{diffLL}), i.e., the difference in the log-likelihood of the two models),
and the difference in the degrees of freedom in the two models (\code{diffDF}).
}
